third_party/WebKit/LayoutTests/bluetooth/server/disconnect/disconnect-fires-event.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
 'use strict';
 // TODO(ortuno): Write tests to check that "Disconnect" was actually
 // called on the device.
 // http://crbug.com/569716
 promise_test(() => {
   return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}]}))
    .then(device => device.gatt.connect())
    .then(gattServer => {
      let disconnected = eventPromise(gattServer.device, 'gattserverdisconnected')
      gattServer.disconnect();
      return disconnected;
    }).then(event => assert_true(event.bubbles));
 }, 'disconnect() fires the gattserverdisconnected event.');
</script>
